requirements.txt
APScheduler==3.11.0
fastapi==0.115.8
uvicorn==0.34.0
atlassian-python-api==3.41.19
SQLAlchemy==2.0.37
PyYAML==6.0.2
urllib3==1.26.6
pandas==2.2.3
openpyxl==3.1.5
pydantic>=2.0
pydantic-settings>=2.0